+++ src/main/java/kitchenpos/domain/OrderLineItem.java
@@ -7,6 +7,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import kitchenpos.menu.domain.Menu;
 
 @Entity
 public class OrderLineItem {
