+++ src/main/java/kitchenpos/menu/service/MenuMapper.java
@@ -1,0 +1,9 @@
+package kitchenpos.menu.service;
+
+import kitchenpos.menu.domain.Menu;
+import org.springframework.stereotype.Component;
+
+@Component
+public class MenuMapper {
+
+}
