+++ src/test/java/kitchenpos/steps/IntegrationTestSteps.java
@@ -5,7 +5,7 @@ import static io.restassured.RestAssured.given;
 import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.dto.MenuDto;
+import kitchenpos.menu.service.MenuDto;
 import kitchenpos.menugroup.service.MenuGroupDto;
 import kitchenpos.dto.OrderDto;
 import kitchenpos.dto.OrderTableDto;
